requirements.txt
Django==1.11.3
psycopg2==2.7.1
social-auth-app-django==1.2.0
django-internationalflavor==0.3.1
social-auth-core==1.4.0
gunicorn==19.7.1
dj-database-url==0.4.2
whitenoise==3.3.0